Pass SHA-1s around as 20 byte arrays instead of 40 (double-wide) character strings.
Summary:
The goal of this revision is (1) to make `Sha1HashCode` take up less memory,
and (2) to make it more efficient to create one of these from data sent over the wire,
which will be important for some upcoming work.

Because I want to have finer-grained control over how `Sha1HashCode` is constructed,
I removed `AbstractSha1HashCode` because `Sha1HashCode` is no longer a product of
a `BuckStyleImmutable`.

There are now three static create methods for `Sha1HashCode`:

* `Sha1HashCode.fromBytes(byte[])`
* `Sha1HashCode.fromHashCode(HashCode)`
* `Sha1HashCode.of(String)`

(I'll rename `Sha1HashCode.of(String)` to `Sha1HashCode.fromString(String)` in a follow-up
revision for consistency, but I didn't want to make this revision any noisier than it already is.)

This also refactors `Sha1HashCode` so it is defined in its own `java_library()` rule.
Ultimately, this will make it easier for `ProjectFilesystem` to take a dependency on it, if appropriate.
(In particular, I'm interested in changing `ProjectFilesystem#computeSha1()` to return a `Sha1HashCode`.)

Test Plan:
* Updated `Sha1HashCodeTest`.
* CI

Reviewed By: k21

fb-gh-sync-id: b44f9f3
fbshipit-source-id: b44f9f3