Include process stderr in exception message
Summary:
This diff adds the stderr from the external worker process directly to the exception message.
Before the stderr logging was happening higher up in the stack and using a different mechanism (a console event) which meant that the stderr logs could potentially be very far away from the exception stack trace, which made debugging harder.

Test Plan:
Edited ##nodeModules.js## so that the curl would fail to simulate a failure.
Had buck build a RN bundle and saw that the stack trace contained the stderr from the worker process:
{F60748497}

Also,
  buck test //test/com/facebook/buck/shell:shell

Reviewed By: marcinkosiba

fbshipit-source-id: f8b2efd