ShellStep: don't write directly to console on error
Summary:
When the `IS_SILENT` option isn't sent to the `ProcessExecutor`,
it'll print stderr/stdout directly to the console when the process
exits with a non-zero return code, which freezes the super console.
Since we already post console events from inside the `ShellStep` for
any stderr/stdout, just always set `IS_SILENT`.

Test Plan:
Ran a build with a failing `genrule` which also prints some output
and verified it didn't freeze the super console.

Reviewed By: beefon

fbshipit-source-id: 27bbb9c