Don't rotate logs on buck rage.
Summary:
This makes it possible for individual commands to specify what they
want their logging setup to be. This is done by using ST to fill
in values in the bucklogging.properties file. Right now the two
supported values are pattern and count.
Due to the way this is set up the user is still free to override
any of these settings.

Test Plan: manual, CI

Reviewed By: Coneko

fbshipit-source-id: 66737b9