Extract interface for RuleKeyBuilder
Summary:
Create a new interface `RuleKeyObjectSink` that can be used instead
of `RuleKeyBuilder` in `appendToRuleKey` calls. This will let us
create classes similar to `RuleKeyBuilder` but with different
output types than just a `RuleKey`, which will be useful for some
input based rule keys.

This diff also changes the output type of `appendToRuleKey` from
`RuleKeyBuilder` to `void` because it was originally likely meant to
enable function call chaining, but that turned out to be impractical
since the return type is not the same as the class that owns the
function.

Test Plan: `buck test`

Reviewed By: Coneko

fbshipit-source-id: a6f3ee6