Show absolute manifest file name when logging errors
Summary:
`ManifestMerger` logs errors with a name of the manifest file and it's difficult to debug especially in a big app when all the files have the same name.

This diff changes logging to show the absolute file name.

Test Plan:
Build an apk with multiple manifests, at least one manifest should not contain the `package` element. Buck should show a message in the following format:

  [/..full/path/to/manifest.xml:1] Missing 'package' attribute in manifest.

Reviewed By: asp2insp

fbshipit-source-id: 2f3f64b