haskell: compile haskell libraries as packages
Summary:
Previously, we'd compile them into pairs of interface files and
libraries.  Switch to building GHC packages to support things like
`--make` and hiding transitive packages.

For now, we only include interface files in the package, so that
we can reuse the same package for both shared and static-pic
compilation.

Test Plan: Added a test that transitive deps aren't included in compilation.

Reviewed By: yiding

fbshipit-source-id: 87c3828