C/C++: don't expect source file to always be first item in dep file
Summary:
In most cases this is true, but when using flags like
`-fsanitize-blacklist=<file>`, `<file>` will show up before the dep
file.  This changes the header verification code to drop paths up until
the input source in the dep file so that finds the header paths correctly.

Test Plan: CI

Reviewed By: luciang

fbshipit-source-id: 045d3bb