lua: avoid race condition when using `PackagedResource`
Summary:
The `PackagedResource` implementation generates the resource in a
non-atomic way, which means that there should only ever be one
`PackagedResource` per cell, otherwise they can race with one another
and cause internal Buck failures.  However, since the constructor for
`PackagedResource` requires a the `ProjectFilesystem` of the relevant
cell, it can't be made static, and so makes it difficult to enforce
this deduplication without attaching more stateful structures to the
descriptions and/or action graph.

This diff switches the lua rules to just use a `WriteFile` rule to
unpack the resources to avoid this race condition.

Test Plan: CI

Reviewed By: beefon

fbshipit-source-id: b3c99ea