Fix linking a dynamic framework that wrapped into a resource
Summary:
There's a case when we build and sign a dynamic framework with Buck and then wrap the framework into a resource to have it in 'Copy Resources' build phase.
Actually it doesn't work this way because Buck fetches all transitive frameworks for target and put these frameworks to 'Copy Files' phase in the target.
It doesn't work even if a framework wrapped into a resource.
This diff fixes that case.

Test Plan:
* Run all unit tests
* Run new test `testAppBundleDoesntLinkFrameworkWrappedWithResource()`

Reviewed By: Coneko

fbshipit-source-id: 48c001b