Cache preprocessor command hashes when generating compilation rules with PCH
Summary:
Previously we were recomputing the preprocessor command and hash for each file
in a target, even though the preprocessor command tends to stay the same. This
was actually a big slowdown.

Here I associate the PreprocessorDelegate with the cache (but avoid putting the
cache on the PreprocessorDelgate itself, since it's persisted) for the duration
of a target's compile rule generation.

This noticibly reduces the the time BUCK spend generating the action graph for
targets that have many files and use PCH.

Test Plan:
Compare action graph generation time before and after.
Use profiler to see that this cost center no longer exist.

Reviewed By: Coneko

fbshipit-source-id: f87a6bd