Make `buck build test/...` work.
Summary:
As noted in the new comment in the code, there was a dependency that `buck autodeps`
could not see because it was part of a generated file, so we had to add it manually.

The reason we did not see any failures in CI is because the dependency was part of
a `java_library()` that powered a benchmark, and apparently this benchmark code is
only run when someone is interested in it, so it did not appear to be built by CI.

I added logic in our `travisci_run.sh` script to ensure this gets run in the future.

Test Plan: Verify that `buck build test/...` failed before this change, but now succeeds.

Reviewed By: andrewjcg

fbshipit-source-id: 00fe264