go: fixes for `exported_deps` and generated sources
Summary:
The change to support the `exported_deps` parameter broke rules that
used generated sources.  This diff fixes this so that dep walks to
find `GoLinkable`s only consider declared and exported deps.

It turns out the exported deps tests were also broken, and as the tests
were running a noop target, this was never caught in the original diff.
This diff also includes fixes to get those working.

Test Plan: added test

Reviewed By: azhavnerchik

fbshipit-source-id: 5be7612