Make ProjectWorkspace constructor package-private.
Summary:
All callers should go through one of:

* `ProjectWorkspace.cloneExistingWorkspaceIntoNewFolder()`.
* `TestDataHelper.createProjectWorkspaceForScenario()`.

In a subsequent revision, I want to introduce some new invariants when
setting up a workspace, so I want to make sure all instances of `ProjectFilesystem`
are created in a consistent way.

In particular, `AndroidBinaryIntegrationTest` and `AndroidExopackageBinaryIntegrationTest`
were using the constructor of `ProjectFilesystem` directly for non-obvious reasons, so
I created `TestDataHelper.createProjectWorkspaceForScenario()` to make the motivation behind
that usage more explicit. I also changed the parameter types to help ensure it is used only
for very specific use cases.

I introduced an interface, `TemporaryRoot`, to unify some uses of `[Debuggable]TemporaryFolder`
and `TemporaryPaths`. This made it possible to delete some code.

Further, the introduction of `TemporaryRoot` made it possible to create an anonymous one in
`InterCellIntegrationTest`, which made it easier to have it use
`ProjectWorkspace.cloneExistingWorkspaceIntoNewFolder()` in favor of the `ProjectWorkspace`
constructor.

Note that I had to move `ProjectWorkspaceTest` into the same package as `ProjectWorkspace`
so that I could make the constructor package-private and `VisibleForTesting` rather than
`public`.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: 3da0d79