Store more keys in metadata
Summary:
We currently store the normal, input-based, and ABI-based rule keys in metadata any
time they are computed. This diff makes us also store the dep-file rule key and the
manifest key if they are computed.

Storing the keys in metadata can be useful when debugging cache poisoning issues.
What can happen is the issue is detected when some build gets a normal or
input-based key hit, but the poisoning happened when an earlier build got an
incorrect manifest hit and pushed the resulting artifact to the other keys.
Seeing the manifest and dep-file keys in the artifact helps track back to
that earlier build.

Test Plan: CI

Reviewed By: marcinkosiba

fbshipit-source-id: ca42dd6