Alpha-sort entries in .classpath file.
Test Plan:
Unfortunately, things still do not build in the latest Eclipse release candidate
(I'm testing in Neon RC2) because of a bad interaction between its annotation processor
and immutables. It's reaaaaly close in that most of the generated code is identical to what
is produced when the annotation processor is run with `javac`, but it is missing `import` statements
for the other types that are generated by the annotation processor.

Reviewed By: Coneko

fbshipit-source-id: c2fdc90