CxxBinary should use platform-free build target for preprocessor input lookup
Summary:
CxxPreprocessables implicitly requires that the incoming build param upon which
to do the traversal is free of platform flavors, as it adds its own.  CxxBinary
rules usually are instantiated with a platform flavor, so this one keeps tracks
of the platform-free target, without passing through the set of all known
platforms to the rule itself, which seems a tad excessive.

Test Plan: A future diff makes use of this fix and includes a integration test.

Reviewed By: Coneko

fbshipit-source-id: 18747f0