Code Sign Command, and preserving code signature after absolutifying the binary
Summary:
Code Signature appears to be at the end of the binary.
During absolutification of the object file paths, we expand string table. Doing it carelessly will override code signature.
This diff fixes this issue - it preserves the code signature, and restores it after string table expansion completes.
Of course it becomes broken after this, but at least it allows to re-sign the binary by reading it and taking entitlements and authority/team ID and code signing on top of the binary.

Test Plan: added tests

Reviewed By: Coneko

fbshipit-source-id: bc01b92