fix external apple packager with default platform not passing SDKROOT
Summary:
We need to enumerate all commands at ImplicitDepsInferringDescription time
(i.e. Target node generation) for possible build targets, since that's the
point that BUCK files are read in.

On top of that, I edited the integraiton test to check that the SDKROOT macro
is actually passed down.

Test Plan: CI

Reviewed By: marcinkosiba

fbshipit-source-id: b63b203