Remove uses of String.format
Summary:
String.format is slow and memory hungry because it creates a formatter
and complex logic for format a String. Where a String is being constructed
from a known set of strings without the need for formatting arguments
it is more efficent to use the + or a StringBuilder.

Test Plan: - Local testing

Reviewed By: beefon

fbshipit-source-id: 3d94731