CommandMode for Main
Summary:
Introduction of CommandMode for `runMainWithExitCode`. This will provide a way to enable/disable
features or logging based on what type the build is. In the future, we can make this readable from the
python part so we can define more modes.

The naming `CommandMode` was because we already have `BuildMode`.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: 21156fb