Fix indentation for instrumentation test outputs
Summary:
RemoteAndroidTestRunner trims whitespace from all the shell
output lines, which trims off indentation on continuation lines.

Going through ddmlib's code, it appeared that they have the ability to
not trim the lines, but that's not exposed through
RemoteAndroidTestRunner, so I had to do some reflection gymnastics to
get hold of it.

But then, there was no nice point to plug in the reflection, so I
created a new test listener to do it.

Test Plan:
unit test to verify that the reflection works. Manually
tested against fbandroid to verify the listener integration.

fbshipit-source-id: 282332c