ignore the executable itself when hashing preprocessor command
Summary:
Previously, `hashCommand` will emit different values if the executable for the
preprocess (or preprocess-and-compile) command are in different locations on
disk. This causes the rule keys to differ for rules using precompiled headers,
since rule keys hash the target names of the dependencies. This reduces the
effectiveness of distributed caches, especially if the writer has the toolchain
in a different directory.

This change is safe because compiler used to generate the PCH is part of the
rule key, so if we get a name collision, the rulekeys would mismatch. As we
don't allow different compilers in the same build, this would simply overwrite
any old pch files without problems. Since PCH is uncachable anyways, this
wouldn't affect caching.

Test Plan:
buck targets --show-rulekey '//:lib#static'

and do it again with a toolchain in a different location, see that they give identical results.

Reviewed By: Coneko

fbshipit-source-id: 609d904