Automatically exclude ignored files from globs.
Summary:
Files that match the ignore path patterns are ignored by many
parts of buck, but not globs. While those files are ignored if
the globs are passed directly into most targets, if they are
used within the build files themselves, for example to generate
a genrule's command string, they can cause issues.
This passes the list of ignored path patterns to the parser so
that globs can ignore them automatically.

Test Plan:
CI
new tests

Reviewed By: ruibm

fbshipit-source-id: 16de5a5