remove_classes for spooling to the disk
Summary:
This enables remove_classes for java_library when spooling is set to the disk. It actually
creates the classes and saves them to the disk but it does not write them to the final Jar.
Again, the matching is happening based on a Pattern. E.g. to remove com.example.Main
you just add `remove_classes = [ 'com.example.Main' ].

JarDirectoryStepHelper already had a Pattern blacklist and I'm re-using that.

Test Plan:
The matching is tested for classes inside a package, stand-alone and member classes.

* New test for JarDirectoryStep: Check that Patterns under remove_classes are deleted as expected.
* New integration test: Make sure final Jar does not contain member, under package or single classes if they match a Pattern.
* CI

Reviewed By: marcinkosiba

fbshipit-source-id: e72d522