Resolver to support JSON configuration and user-pass authentication
Summary:
Ref #664

Modifies `Resolver` to support:

  * basic authentication
  * multiple Maven repositories at the same time
  * reading repository, Maven, and artifact settings from a file
  * optionally overridding these from the command line.

The PR a bit wobbly, as some of the `Resolver` code is now only used in the integration tests. Will fix, but wanted to make sure this is something you guys are ok having merged :D
Closes https://github.com/facebook/buck/pull/691

Test Plan:
Haven't touched the unit tests yet, but tried it out on some artifacts
I needed to download.

Reviewed By: Coneko

fbshipit-source-id: 8a1982a