Remove warning message during dx that affects buck build progress printing
Summary:
When using libraries like leakcanary dx prints warnings like
```
warning: Ignoring InnerClasses attribute for an anonymous inner class
(com.squareup.haha.guava.base.Joiner$1) that doesn't come with an
associated EnclosingMethod attribute. This class was probably produced by a
compiler that did not target the modern .class file format. The recommended
solution is to recompile the class from source, using an up-to-date compiler
and without specifying any "-target" type options. The consequence of ignoring
this warning is that reflective operations on this class will incorrectly
indicate that it is *not* an inner class.
```

That completely stalls the buck build progress from being output to stdout. The build will silently pass/fail without any indication whatsoever due to this after being stuck

```
[+] BUILDING...5.8s [28%] (227/495 JOBS, 227 UPDATED, 227 [45.9%] CACHE MISS)
```

This change prevents the warning from being print to the console.
Closes https://github.com/facebook/buck/pull/752

Reviewed By: Coneko, dreiss

fbshipit-source-id: 0b9cda7