Introduce `buck suggest`.
Summary:
This introduces a new command, `buck suggest`, that takes the build target
for a `java_library()` rule and attempts to print out a repartitioning of the rule
where each set of `srcs` is as small as possible. This is intended to aid
developers into partitioning large `java_library()` rules into more fine-grained ones,
which should result in faster incremental build times because fewer unaffected `.java`
files will have to be recompiled as a result.

This is achieved by using the `JavaDepsFinder` to find the dependency relationships
within a set of `srcs` and then applying Tarjan's strongly connected components
algorithm on the resulting DAG. Each strongly connected component that is produced
from the algorithm becomes an individual `java_library()` rule in the output.

If everything works, it should be possible to replace the original definition for the
`java_library()` with the output from `buck suggest`. This is
possible because it produces a rule with the same name as the original with all
of the new, fine-grained rules as its `exported_deps`. In practice, to achieve better
performance, developers would phase out the use of the original, "fat" rule, with
uses of the appropriate fine-grained rules. If the other rules in the project are
defined with `autodeps = True`, then this should be able to be done automatically by Buck.
(This scenario is verified by the integration test.)

Test Plan:
Created an integration test: `SuggestPartitionIntegrationTest`.
Previewed updated docs locally in Chrome.

Reviewed By: marcinkosiba

fbshipit-source-id: d5878aa