Switch Memoization to SoftRef
Summary:
Memoizing getArchiveMembers put a huge memory burden on builds (33% of the heap!) because
of the following retention path:
`Build->DefaultJavaLibrary->JarArchiveDependencySupplier->archiveMembers`
This means that every single `Path` object to every single class file is retained for the full
duration of the build.

This change moves to a soft reference, allowing the memoized set to be reclaimed. This led
to a reduction in the number of cache hits (80% cache hit instead of 83%) but eliminated the
extra memory pressure entirely.

Test Plan:
Tested both on SSD and Spinning Disk. On spinning disk, the performance difference now
for a full fb4a build is around 13 minutes of CPU time (amortized this is much smaller)
between the old input based rule keys and the new dependency based rule keys.
The build of FB4A no longer OOMs when it's 1/3 of the way done.

Reviewed By: jkeljo

fbshipit-source-id: 008d5fe