Cache getBasePathAncestorTarget result
Summary:
getBasePathAncestorTarget was taking ~10% of memory under some conditions
because it was being called for many target files which shared a common
ancestor.

This introduces a cache at one level above the initial file which keeps
the cache size small, but can halve the amount of memory used.

Test Plan: - Local testing

Reviewed By: Coneko

fbshipit-source-id: 34939dd