Replace an empty buck target with //...
Summary: An empty target wasn't matching all targets for a sliceless project because it was looking for a project_config option, so, when using the new project generator and not specifying a slice, we should treat it as if the user has specified //... (i.e. everything).

Test Plan:
Local testing
Sandcastle.

Reviewed By: cotizo

fbshipit-source-id: 751b1b3