Share (most) strip rules between Android apps
Summary:
Previously, rules for stripping libs for inclusion in Android apps used
the app target as the base for the strip target.  This prevented them
from being shared between apps (even slight variations like exo/noexo),
so someone building an app that wasn't pre-cached (the exact same
target) would get a cache miss on the strip rule and have to download
the cache artifact for the unstripped lib, which can be very large.

Now, use the library target as the base of the target to allow sharing.

For libs that are not BuildTargetSourcePaths, keep the old behavior to
be safe.  I've only seen this affect the C++ runtime, whose unstripped
lib is part of the NDK.

Test Plan: Added integration test.

Reviewed By: andrewjcg, asp2insp

fbshipit-source-id: 098e4da