Distributed build TargetGraph serialization
Summary:
This adds the ability for the distributed build command to serialize and deserialize the TargetGraph.
The deserialization is a bit broken because it ends up using the rootCell for the freshly instantiated
TargetNode's deps. I plan to address this in a later change.

Test Plan: CI, added new test

Reviewed By: Coneko

fbshipit-source-id: ff2762c