Fixed non-deterministic missing cache uploads issue
Summary:
The two-level cache was previously using the same executor service as the http/thrift cache uses to upload artifacts. This meant that if the two-level cache had scheduled Futures that were yet to be processed on said executor by the time the build completes, those Futures would never be able to subsequently schedule the actual uploads to the remote cache (as once the build is complete the executor service is locked for new work).

The new behavior is that the two-level cache uses a direct executor, meaning artifacts are scheduled to be uploaded to the remote cache on the same threads as the CachingBuildEngine runs the build itself (i.e. by the time the build is complete, all artifacts should be scheduled for upload).

Test Plan:
ci
'buck build buck' multiple times and make sure full artifact count is uploaded

Reviewed By: marcinkosiba

fbshipit-source-id: 072d031