Do not allow "repositories" override
Summary: For configuration overrides, we want to move to a cell path to rawConfigs map. However, to build the raw configs, we need to parse the configuration overrides. This introduces a circular dependency where configuration parsing requires knowing all of the cells, and cell creation requires parsing the configuration. This breaks the dependency by fiat: no cell overrides from the command line.

Test Plan: Add a unit test

Reviewed By: marcinkosiba

fbshipit-source-id: f9c7e1a