MoreFiles.deleteRecursivelyWithOptions(): add new `ErrorHandler` class
Summary:
This diff adds a new `ErrorHandler` class that the caller can pass in
to `MoreFiles.deleteRecursivelyWithOptions()` to customize how errors
are handled when they are encountered.  Default handlers are added for
throwing all errors, ignoring missing files, and only warning on errors.

This replaces the `IGNORE_NO_SUCH_FILE_EXCEPTION` option which would
previously silently abort the recursive delete on the first
`NoSuchFileException` thrown.  Now, the delete will be completed even
in the event of several `NoSuchFileException` exceptions.

This also changes the `AsynchronousDirectoryContentsCleaner` so that
it log each error while still completing the delete, rather than aborting
the delete on the first error and logging it.

Test Plan: CI

Reviewed By: beefon

fbshipit-source-id: 5e2a768