Let ShellStep throw IOException
Summary:
When `ShellStep` fails because it cannot execute the file specified
in the command, it exits with an unhelpful error message stating
that the command exited with code 1. This change instead makes
`ShellStep` propagate the exception outside of the step so the
error message can be reported to the user. This also makes it
easier for its subclasses to handle `IOException`s.

Test Plan: `buck test`

Reviewed By: marcinkosiba

fbshipit-source-id: 82fb5a2