Improve ZipOutputStreamTest
Summary:
Some unit tests that test STORE mode were not setting the entry size. This would produce incorrect output, but was not caught because it was not tested for. Note that this was just a slight incompleteness on the unit test side, the code itself is fine (or we hope so).
Another issue with unit tests was that not all of the tests were applied to all of the HandleDuplicate modes. And in cases where they were, the code would be copy-pasted with only a single constant being different.
This diff fixes both of the issues.

Test Plan: buck test

Reviewed By: Coneko

fbshipit-source-id: 0a699a9