Use THROW_EXCEPTION instead of OVERWRITE_EXISTING in ZipStep
Summary: It seems that ZipStep doesn't have to use the inefficient OVERWRITE_EXISTING mode. OverwritingZipOutputStream uses ZipEntry.getName() to identify entries, and so does ZipStep when building a TreeMap. That means that ZipEntries in the TreeMap are unique and we can just use the more efficient THROW_EXCEPTION mode.

Test Plan:
build and run fb4a and messenger from scratch and with --no-cache option

  ~/buck/bin/buck clean
  NO_BUCKD=1 ~/buck/bin/buck install --no-cache messenger

Reviewed By: marcinkosiba

fbshipit-source-id: f7690c0