lua: minor binary description refactoring
Summary:
This just refactors the code in `LuaBinaryDescription`, so that
we generate the Lua binary starter in the same place where we collect
up the Lua components from transitive deps.  This is a precursor to
another diff which adds omnibus linking support for Lua rules, which
will require making component gathering and starter generation happen
together.

Test Plan: CI

Reviewed By: beefon

fbshipit-source-id: 9ea2547