Refactor environment handling in ProjectWorkspace.
Summary:
All the callsites are either passing absent or a map built starting
from the system environment, so we can refactor it to add the system
environment in the method itself. This has the advantaged the system
environment gets filtered in the same way as when no overrides are
specified.

Test Plan: CI

Reviewed By: marcinkosiba

fbshipit-source-id: 8a7d4cc