Console parsing status: Don't print "Watchman overflow" on first build
Summary:
In `Expose parsing status to console`, michath added a nice
feature to let users know why parsing was slow.

Previously, the design of `WatchmanWatcher` always posted an overflow
event on the very first `buckd` command. This was because a new buckd
instance would use a fresh UUID in the `since` query, so `watchman`
would always return `is_fresh_instance = true` in the first query
result.

This used to be harmless, but after `Expose parsing status to
console`, we would always print `Watchman overflow` for the first
command issued to `buckd`.

With this diff, I change `WatchmanWatcher` to only post an
overflow event if a subsequent command after the first indicates
`is_fresh_instance = true` (say, if `watchman` has to recrawl the
watch for any reason).

Test Plan:
Unit tests updated.
  ant java-test -Dtest.class=WatchmanWatcherTest
  ant java-test -Dtest.class=WatchmanWatcherIntegrationTest
  ant java-test -Dtest.class=DaemonIntegrationTest

Reviewed By: michath

fbshipit-source-id: b80b50a