Improve efficiency of set operations for DepFiles
Summary:
When profiling dep file based rule keys for java, I found 29.64% of my CPU
time was taken up by `ManifestKeyBuilder.build()` which spends all its time in 2 calls to `ImmutableSet.copyOf()`.

The first call can be eliminated since the target requires an `Iterable` and the second call can be sped up.
The reason the `copyOf` operation is so expensive is not actually the copying. It's the iteration over the
`SetView` instances returned from `Sets.difference` and `Sets.intersection`. These construct `Predicate` instances
which check for membership in the second set. The second set in these calls, however, is a `SortedImmutableSet` which
implements `contains()` via binary search. This means `O(log(n))` runtime instead of `O(1)` and even worse,
instead of calling `hashCode` (which is cached and O(1)), we call `compare` which does a very expensive `Path` equality
comparison.

Copying the set before taking the intersection converts it to a hash-based set with the desired `O(1)` runtime for `contains`

Test Plan:
All related unit tests pass. I ran the test again and found that calls to this function have dropped to < 1% of CPU time
and the total time taken for the build (13488 build nodes) dropped by 10%.

Reviewed By: mkillianey

fbshipit-source-id: fd7310c