Avoid checking for existence of x/BUCK when x is likely to be a file
Summary:
A small optimization for buck package boundary checking.

The initial path passed into getBasePathOfAncestorTarget is likely to be a file
(since this is called on all source files). It's therefore advantageous to
check this fact, and don't try to check for x/BUCK since files are not
directories. This avoids adding obvious non-existent files into the path
existence cache, and likely plays better with OS filesystem caches, as it would
avoid cache misses in the VFS name cache and filling up the said cache with
negative vnode entries.

Also add some doc comments to existing things in the file.

Test Plan:
Profile a buck invocation which loads a large target.

  NO_BUCKD=1 \
  BUCK_EXTRA_JAVA_ARGS="-agentpath:/path/to/libyjpagent.jnilib=sampling,onexit=snapshot" \
  ~/buck/bin/buck targets --show-rulekey some-target

Reviewed By: alsutton

fbshipit-source-id: 6454de6