Move computeSha1() from ProjectFilesystem to ProjectFilesystemDelegate.
Summary:
As part of this move, I changed the implementation so that rather than loading
the file into memory as an array of bytes and then hashing it, it does it in a
streaming fashion by leveraging Guava's `Files.hash()` method.

This method will have a special implementation in `EdenProjectFilesystemDelegate`.

Test Plan: CI

Reviewed By: marcinkosiba

fbshipit-source-id: d8c82b6