Make ProjectFilesystem#getRootPath() final.
Summary:
Some implementations of `FakeProjectFilesystem` were overriding `getRootPath()`,
but the main reason they had to do this was to avoid the constructor checks
in `ProjectFilesystem` that verified that the root path actually exists. This is
frequently unnecessary when creating a `FakeProjectFilesystem`, so this
introduces a new API that allows subclasses of `ProjectFilesystem` to disable
these checks.

This makes it possible to eliminate a number of anonymous subclasses of
`ProjectFilesystem`.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: 080d4f9