Remove ProjectFilesystem#resolvePathFromOtherVfs().
Summary:
It appears that `getPathForRelativePath()` was just a passthrough for
`resolvePathFromOtherVfs()`. Also, it was `protected` yet never overridden
or accessed from a subclass.

Note that I renamed the `pathRelativeToProjectRootOrJustAbsolute` parameter to
`pathRelativeToProjectRoot` because it is always passed to `resolve()`, so it
seems like it should always be a relative path.

This is needed for an upcoming refactoring for `ProjectFilesystemDelegate`.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: 3974e02