ProjectFilesystemDelegateFactory now produces an EdenProjectFilesystemDelegate, if appropriate.
Summary:
This is the first step in true Buck/Eden integration. If Buck detects that it is
building within an instance of EdenFS, then it will create an
`EdenProjectFilesystemDelegate` instead of a `DefaultProjectFilesystemDelegate`.

Note how the `EdenProjectFilesystemDelegate` takes advantage of a special Eden
API to implement `computeSha1()`, which should be more efficient than the
default implementation of reading the file and hashing the bytes in Java.

Test Plan: TBD

Reviewed By: marcinkosiba

fbshipit-source-id: a486829