Log even fewer perf events when parsing.
Summary:
This introduces a minimum duration perf event that allows us to
further decrease the size of the .trace files.
The new scope will only log the event if it lasts for longer than
a specified time, otherwise it will be merged into the parent event.

Test Plan: CI, manual

Reviewed By: Coneko

fbshipit-source-id: 82cf289