lua: add initial omnibus linking support
Summary:
This adds support for using omnibus linking for native code pulled
in by Lua binary rules (just as it currently works for Python).
For now, we exclude the native starter and it's deps from the omnibus
link, as there are still a few more changes needed to supoprt this.

Test Plan: added tests

Reviewed By: beefon

fbshipit-source-id: 6ee161a