Stop propagating ExecutionException in project generator code paths and swallowing in HumanReadableException
Summary:
API which do not deal with the concept of execution should not throw
ExecutionExceptions, instead the cause should be thrown instead (or
dealt with in some other way).

ExecutionExceptions should definitely not be wrapped in HumanReadableExceptions wholesale, as the causes themselves may not be HumanReadableExceptions.

This was causing certain errors in project generation to be hidden behind a common failure message, making it hard to debug in the field.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: 6d5230b