Add prebuilt dynamic framework support
Summary:
Manually imported from https://github.com/facebook/buck/pull/818

This PR proposes a way to embed prebuilt dynamic frameworks into `apple_bundle`s, linking that framework against the binary for the apple bundle. This is a start to solving #440, in buck it (mostly) works, the project generation still has to be done. The rule syntax for this would be:
```
prebuilt_apple_framework(
    name = 'BuckTest',
    framework = 'BuckTest.framework',
)
```

You can add this rule to an apple_library or apple_binary, this should do the following:
- Symlink the `Headers` folder inside the framework as the exposed headers for the rule.
- Export linkerflag `-rpath loader_path/Frameworks`/ `-rpath executable_path/Frameworks`
- Export the linkerflag `-framework <frameworkname>`, and `-F path/to/framework/`
- Embed the framework in the bundle's "Frameworks" folder

I've updated the PR to do the following:

- In cxx preprocess deps gathering step, go through the frameworks and see if any is a buildrule. If so add that rule as a preprocessing dep step.
- Simplify the `prebuilt_apple_framework` build rule to copy the framework to a gen folder, and pass the framework search paths option`-F` to clang on linking and as an preprocessor dep.

Lets keep this PR restricted to just the prebuilt framework, and I'll incorporate the feedback for frameworks from `apple_library`'s (and removing the apple library #framework flavor) in a future one.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: 772e3c4