lua: fix packaging style target/output inconsistency
Summary:
Previously, the build target name referring to the rule which generated
the Lua binary output changed depending on the package style.  This meant
that switching between package styles could cause the output to get
overwritten without invalidating the metadata for the build target using
the other package style.  Then when a subsequent build switched pacakge
modes again, the output path would not be updated, resulting in stale
outputs.

This diff fixes the Lua binary code to always use the same build target
for generating a given output.

Test Plan: Added an integration test which triggered the bug.

Reviewed By: adamlerer

fbshipit-source-id: ddddc3d