introduce broadcastEvent and listener
Summary:
introducing BroadcastEventListener. A listener that makes sure to broadcast an event to all the subscribed buses.
the reasoning behind it is that it's possible some of the events should be received from whoever is running a
command at the time since it might affect that command too. In case you run multiple commands at the same
time and one invalidates caches the other command should know.

the other case that it solves is when we need to communicate information between parts of the daemon and the
current invocation. BroadcastEventListener generates an abstraction that allows us to send the events to the
subscribed buses. the listener throws if you try to send an event when there is no subscribed listener.

introducing interface BroadcastEvent to help us track which events are supposed to be broadcasted and what
should not.

Test Plan:
* a test that throws when there is no bus to receive the events.
* a test that counts that all events arrive when you have multiple events.
* manual testing that the newly switched ActionGraphCache events are still being delivered.

Reviewed By: marcinkosiba

fbshipit-source-id: b72a09a